Remove unused controller code
