+++ operator/pkg/helmreconciler/render.go
@@ -452,3 +452,21 @@ func (h *HelmReconciler) getCRHash(componentName string) (string, error) {
 	}
 	return strings.Join([]string{crName, crNamespace, componentName}, "-"), nil
 }
+
+// ChartManifestsMap is a typedef representing a map of chart-name: []manifest, i.e. the manifests
+// associated with a specific chart
+type ChartManifestsMap map[string][]releaseutil.Manifest
+
+// Consolidated returns a representation of mm where all manifests in the slice under a key are combined into a single
+// manifest.
+func (mm ChartManifestsMap) Consolidated() map[string]string {
+	out := make(map[string]string)
+	for cname, ms := range mm {
+		allM := ""
+		for _, m := range ms {
+			allM += m.Content + helm.YAMLSeparator
+		}
+		out[cname] = allM
+	}
+	return out
+}
